# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [account]
      summary: Get the current authenticated user
      description: >
        Returns the current user, provisioning a local record from the auth
        provider on first call. Includes profile, verification, and activation
        state so the client can route the user into onboarding, verification,
        or the app.
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [account]
      summary: Permanently delete the current user's account
      description: >
        Irreversibly deletes the authenticated user's auth-provider identity and
        all of their data (profile, predictions, owned challenges, memberships,
        chat messages, social graph). Satisfies Apple App Store Guideline
        5.1.1(v). The client must sign out after a successful response.
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Auth provider deletion failed; nothing was deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/profile:
    patch:
      operationId: updateProfile
      tags: [account]
      summary: Update the current user's profile
      description: >
        Update private real name, public display name, unique @username, and
        avatar. Display name and username are validated for platform-wide
        uniqueness.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Display name or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/display-name-availability:
    get:
      operationId: checkDisplayNameAvailability
      tags: [account]
      summary: Check if a public display name is available
      parameters:
        - name: displayName
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResult"
  /me/username-availability:
    get:
      operationId: checkUsernameAvailability
      tags: [account]
      summary: Check if a @username handle is available
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResult"
...
  /me/favorite-team:
    patch:
      operationId: updateFavoriteTeam
      tags: [account]
      summary: Set or change the user's favourite team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId]
              properties:
                teamId:
                  type: string
                  description: UUID of the team to select
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/favorite-club:
    patch:
      operationId: updateFavoriteClub
      tags: [account]
      summary: Set or change the user's favourite club
      description: >-
        Sets the user's favourite CLUB. Optional and additive to the national
        favourite team (never participates in activation). The target team must
        be a club; national teams are rejected with 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamId]
              properties:
                teamId:
                  type: string
                  description: UUID of the club to select
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Club not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: >
            Optional machine-readable error code for client-side branching
            (e.g. "owner_pool_full" when the host's shared participant pool
            is exhausted).
      required:
        - error
    CurrentUser:
      type: object
      description: The current authenticated user with profile and activation state.
      properties:
        id:
          type: string
          description: Internal user id (uuid)
        email:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        realName:
          type: ["string", "null"]
          description: Private real name, never shown publicly
        displayName:
          type: ["string", "null"]
          description: Unique public display name
        username:
          type: ["string", "null"]
          description: Unique @username handle
        avatarUrl:
          type: ["string", "null"]
        mobileNumber:
          type: ["string", "null"]
        mobileVerified:
          type: boolean
        role:
          type: string
          description: user or admin
        level:
          type: string
          description: Gamification level (bronze, silver, gold, elite, legend)
        totalPoints:
          type: integer
        profileComplete:
          type: boolean
          description: True when display name and username are set
        favoriteTeamSelected:
          type: boolean
          description: True when the user has chosen a favourite team
        favoriteTeam:
          oneOf:
            - $ref: "#/components/schemas/FavoriteTeamRef"
            - type: "null"
        favoriteClubSelected:
          type: boolean
          description: >
            True when the user has chosen a favourite club. Optional and additive
            to the national favourite team; never participates in activation.
        favoriteClub:
          oneOf:
            - $ref: "#/components/schemas/FavoriteClubRef"
            - type: "null"
        activated:
          type: boolean
          description: True when email + mobile verified and profile complete
        hidePredictions:
          type: boolean
          description: >
            User-level privacy: when true, the player's recent predictions are
            hidden from everyone else's view of their public profile.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - emailVerified
        - mobileVerified
        - role
        - level
        - totalPoints
        - profileComplete
        - favoriteTeamSelected
        - favoriteClubSelected
        - activated
        - hidePredictions
        - createdAt
    FavoriteTeamRef:
      type: object
      description: Compact team reference for favourite team display.
      properties:
        id:
          type: string
        nameEn:
          type: string
        nameAr:
          type: string
        flagUrl:
          type: ["string", "null"]
      required:
        - id
        - nameEn
        - nameAr
    FavoriteClubRef:
      type: object
      description: Compact club reference for favourite club display.
      properties:
        id:
          type: string
        nameEn:
          type: string
        nameAr:
          type: string
        crestUrl:
          type: ["string", "null"]
      required:
        - id
        - nameEn
        - nameAr
    ProfileUpdate:
      type: object
      description: Editable profile fields. Omit a field to leave it unchanged.
      properties:
        realName:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          type: string
          minLength: 2
          maxLength: 40
        username:
          type: string
          minLength: 3
          maxLength: 30
        avatarUrl:
          type: string
        termsAccepted:
          type: boolean
          description: >
            Set to true when the user has accepted the Terms of Service and
            Privacy Policy. The server records the acceptance timestamp and the
            version of the legal text in force at that time.
    AvailabilityResult:
      type: object
      properties:
        value:
          type: string
        available:
          type: boolean
        reason:
          type: ["string", "null"]
          description: Why it is unavailable (taken, invalid, reserved)
      required:
        - value
        - available
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...