# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FavoriteClubRef } from './favoriteClubRef';
import type { FavoriteTeamRef } from './favoriteTeamRef';

/**
 * The current authenticated user with profile and activation state.
 */
export interface CurrentUser {
  /** Internal user id (uuid) */
  id: string;
  /** @nullable */
  email?: string | null;
  emailVerified: boolean;
  /**
     * Private real name, never shown publicly
     * @nullable
     */
  realName?: string | null;
  /**
     * Unique public display name
     * @nullable
     */
  displayName?: string | null;
  /**
     * Unique @username handle
     * @nullable
     */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  mobileNumber?: string | null;
  mobileVerified: boolean;
  /** user or admin */
  role: string;
  /** Gamification level (bronze, silver, gold, elite, legend) */
  level: string;
  totalPoints: number;
  /** True when display name and username are set */
  profileComplete: boolean;
  /** True when the user has chosen a favourite team */
  favoriteTeamSelected: boolean;
  favoriteTeam?: FavoriteTeamRef | null;
  /** True when the user has chosen a favourite club. Optional and additive to the national favourite team; never participates in activation.
   */
  favoriteClubSelected: boolean;
  favoriteClub?: FavoriteClubRef | null;
  /** True when email + mobile verified and profile complete */
  activated: boolean;
  /** User-level privacy: when true, the player's recent predictions are hidden from everyone else's view of their public profile.
   */
  hidePredictions: boolean;
  createdAt: Date;
}